Removed double run of unit test.
